refactor: utils.tsx -> ts 파일로 수정 / 컴포넌트 리턴 함수 분리
