+++ src/app/(pages)/App.tsx
@@ -1,9 +1,8 @@
 "use client";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import Modal from "@components/atoms/modal/Modal";
+import Modal, { getModal } from "@components/atoms/modal/Modal";
 import useModalStore from "@/_stores/client/modalStore";
-import { getModal } from "@/_utils/utils";
 import { useEffect, useState } from "react";
 
 export default function App({
